Create a UPN service from lengthy code in the custom validator
